setup wireframes to each scene via buttons rather than tabs
